consolidate domain tests and rename
